requirements.txt
langchain==0.3.27 
langchain-community==0.3.27 
langchain-core==0.3.74 
langchain-openai==0.3.30 
langchain-text-splitters==0.3.9
openai==1.99.9
faiss-cpu==1.12.0
fastapi==0.116.1
uvicorn==0.35.0
aiohttp==3.12.15
streamlit==1.48.1
playwright==1.53.0
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.4
